ClearString class is back, reading tag file For hashTags the key is cleaned For user input with tags and movies the input is also cleaned Movie name is cleaned when inserted on Trie updated TODO's

+++ src/Genre.cpp
@@ -21,7 +21,7 @@ vector<Movie*> Genre::getTop(int n){
         isSorted = true;
     }
     
-    for(int i=0; i < n && i < movies.size(); i++){
+    for(int i=0; i < n && i < (int)movies.size(); i++){
         ans.push_back(movies[i]);
     }
